fixes and improvements

+++ scripts/install-essential.sh
@@ -1,15 +1,35 @@
 #!/bin/bash
 
+set -eE -o functrace
+
+clean_up() {
+  local lineno=$1
+  local msg=$2
+  echo "Failed at $lineno: $msg"
+  exit 1
+}
+trap 'clean_up $LINENO "$BASH_COMMAND"' ERR
+
+addLine() {
+  if ! grep -q "$1" "$2" 2>/dev/null; then
+    (echo; echo "$1") >> "$2"
+  fi
+}
+
 # set fish as default shell
-chsh -s /usr/bin/fish "$USER"
+if ! grep -Fxq "/usr/bin/fish" /etc/shells
+then
+  chsh -s /usr/bin/fish "$USER"
+fi
 
 # configure fish keybindings
-mkdir -p "$HOME"/.config/fish/functions
-touch "$HOME"/.config/fish/config.fish
+mkdir -p "$HOME/.config/fish/functions"
+touch "$HOME/.config/fish/config.fish"
 chown -R "$USER":"$USER" "$HOME"
 
 keybindings="$HOME/.config/fish/functions/fish_user_key_bindings.fish"
-tee "$keybindings" <<EOF
+if [ ! -f "$keybindings" ]; then
+  tee "$keybindings" <<EOF
 function fish_user_key_bindings
   #bind \x7F 'backward-kill-bigword'
   bind \e\[3\;3~ delete-current-history-search
@@ -21,29 +41,37 @@ function fish_user_key_bindings
   bind \\Cz undo
 end
 EOF
+fi
 
 # install starship shell prompt
-wget https://starship.rs/install.sh -O install-starship.sh
-chmod +x install-starship.sh
-./install-starship.sh -y
-rm install-starship.sh
-echo 'eval "$(starship init bash)"' >>"$HOME"/.bashrc
-echo 'starship init fish | source' >>"$HOME"/.config/fish/config.fish
+if ! command -v starship &> /dev/null
+then
+  wget https://starship.rs/install.sh -O install-starship.sh
+  chmod +x install-starship.sh
+  ./install-starship.sh -y
+  rm install-starship.sh
+  addLine 'eval "$(starship init bash)"' "$HOME/.bashrc"
+  addLine 'starship init fish | source' "$HOME/.config/fish/config.fish"
+fi
 
 # install homebrew
 chown -R "$USER":"$USER" "$HOME"
 su -P "$USER" -c "USER=$USER HOME=$HOME ./install-essential-user.sh"
 
 # install node with nvs
-export NVS_HOME="$HOME/.nvs"
-echo 'export NVS_HOME="$HOME/.nvs"' >> "$HOME"/.bashrc
-echo 'export NVS_HOME="$HOME/.nvs"' >> "$HOME"/.config/fish/config.fish
-git clone https://github.com/jasongin/nvs "$NVS_HOME"
-. "$NVS_HOME/nvs.sh" install
-[ -s "$NVS_HOME/nvs.sh" ] && . "$NVS_HOME/nvs.sh"
-nvs add 18
-nvs link 18
-nvs use 18
+if [ ! -d "$HOME/.nvs" ]; then
+  export NVS_HOME="$HOME/.nvs"
+  git clone https://github.com/jasongin/nvs "$NVS_HOME"
+  . "$NVS_HOME/nvs.sh" install
+  [ -s "$NVS_HOME/nvs.sh" ] && . "$NVS_HOME/nvs.sh"
+  nvs add 18
+  nvs link 18
+  nvs use 18
+fi
+
+addLine 'export NVS_HOME="$HOME/.nvs"' "$HOME/.bashrc"
+addLine '[ -s "$NVS_HOME/nvs.sh" ] && . "$NVS_HOME/nvs.sh"' "$HOME/.bashrc"
+addLine 'export NVS_HOME="$HOME/.nvs"' "$HOME/.config/fish/config.fish"
 
 # enable docker service iif docker is installed
 if command -v docker; then
@@ -51,8 +79,9 @@ if command -v docker; then
   systemctl enable containerd.service
 
   # configure json-file logging
-  mkdir -p /etc/docker
-  tee /etc/docker/daemon.json <<EOF
+  if ! grep -q '"max-size": "10m"' '/etc/docker/daemon.json' 2>/dev/null; then
+    mkdir -p /etc/docker
+    tee /etc/docker/daemon.json <<EOF
 {
   "log-driver": "json-file",
   "log-opts": {
@@ -61,10 +90,7 @@ if command -v docker; then
   }
 }
 EOF
+  fi
 fi
 
 chown -R "$USER":"$USER" "$HOME"
-
-# cleanup
-apt-get autoremove -y
-apt-get clean
